requirements.txt
Flask==2.3.2
pymongo==4.4.0
dnspython==2.3.0
python-dotenv==1.0.0